refactor: moved skl, ogf operators into packed
