+++ io_scene_xray/plugin.py
@@ -10,7 +10,7 @@ from .ops import BaseOperator as TestReadyOperator
 from .ui import collapsible
 from .utils import (
     AppError, ObjectsInitializer, logger, execute_with_logger,
-    execute_require_filepath, FilenameExtHelper
+    execute_require_filepath, FilenameExtHelper, mk_export_context
 )
 from .xray_motions import MOTIONS_FILTER_ALL
 from . import plugin_prefs
@@ -21,17 +21,8 @@ from .scene import ops as scene_ops
 from .obj.exp import ops as object_exp_ops
 from .obj.imp import ops as object_imp_ops
 from .anm import ops as anm_ops
-
-
-def invoke_require_armature(func):
-    def wrapper(self, context, event):
-        active = context.active_object
-        if (not active) or (active.type != 'ARMATURE'):
-            self.report({'ERROR'}, 'Active is not armature')
-            return {'CANCELLED'}
-        return func(self, context, event)
-
-    return wrapper
+from .skl import ops as skl_ops
+from .ogf import ops as ogf_ops
 
 
 class BaseSelectMotionsOp(bpy.types.Operator):
@@ -106,222 +97,6 @@ class _MotionsList(bpy.types.UIList):
 
 
 @registry.module_thing
-@registry.requires('Motion')
-class OpImportSkl(TestReadyOperator, io_utils.ImportHelper):
-    class Motion(bpy.types.PropertyGroup):
-        flag = bpy.props.BoolProperty(name='Selected for Import', default=True)
-        name = bpy.props.StringProperty(name='Motion Name')
-
-    bl_idname = 'xray_import.skl'
-    bl_label = 'Import .skl/.skls'
-    bl_description = 'Imports X-Ray skeletal amination'
-    bl_options = {'UNDO'}
-
-    filter_glob = bpy.props.StringProperty(default='*.skl;*.skls', options={'HIDDEN'})
-
-    directory = bpy.props.StringProperty(subtype='DIR_PATH')
-    files = bpy.props.CollectionProperty(type=bpy.types.OperatorFileListElement)
-
-    motions = bpy.props.CollectionProperty(type=Motion, name='Motions Filter')
-    __parsed_file_name = None
-
-    def draw(self, context):
-        layout = self.layout
-        row = layout.row()
-        row.enabled = False
-        row.label('%d items' % len(self.files))
-
-        motions, count = self._get_motions(), 0
-        text = 'Filter Motions'
-        enabled = len(motions) > 1
-        if enabled:
-            text = 'Filter Motions: '
-            count = len([m for m in motions if m.flag])
-            if count == len(motions):
-                text += 'All (%d)' % count
-            else:
-                text += str(count)
-        _, box = collapsible.draw(
-            layout,
-            self.bl_idname,
-            text,
-            enabled=enabled,
-            icon='FILTER' if count < 100 else 'ERROR',
-            style='nobox',
-        )
-        if box:
-            col = box.column(align=True)
-            BaseSelectMotionsOp.set_motions_list(None)
-            col.template_list(
-                '_MotionsList', '',
-                self, 'motions',
-                context.scene.xray.import_skls, 'motion_index',
-            )
-            row = col.row(align=True)
-            BaseSelectMotionsOp.set_data(self)
-            row.operator(_SelectMotionsOp.bl_idname, icon='CHECKBOX_HLT')
-            row.operator(_DeselectMotionsOp.bl_idname, icon='CHECKBOX_DEHLT')
-            row.operator(_DeselectDuplicatedMotionsOp.bl_idname, icon='COPY_ID')
-
-    def _get_motions(self):
-        items = self.motions
-        if len(self.files) == 1:
-            fpath = os.path.join(self.directory, self.files[0].name)
-            if self.__parsed_file_name != fpath:
-                items.clear()
-                for name in self._examine_file(fpath):
-                    items.add().name = name
-                self.__parsed_file_name = fpath
-        else:
-            items.clear()
-        return items
-
-    @staticmethod
-    def _examine_file(fpath):
-        if fpath.lower().endswith('.skls'):
-            from .xray_motions import examine_motions
-            with open(fpath, 'rb') as file:
-                return examine_motions(file.read())
-        return tuple()
-
-    @execute_with_logger
-    def execute(self, context):
-        if not self.files:
-            self.report({'ERROR'}, 'No files selected')
-            return {'CANCELLED'}
-        from .skl.imp import import_skl_file, import_skls_file, ImportContext
-        motions_filter = MOTIONS_FILTER_ALL
-        if self.motions:
-            selected_names = set(m.name for m in self.motions if m.flag)
-            motions_filter = lambda name: name in selected_names
-        import_context = ImportContext(
-            armature=context.active_object,
-            motions_filter=motions_filter,
-        )
-        for file in self.files:
-            ext = os.path.splitext(file.name)[-1].lower()
-            fpath = os.path.join(self.directory, file.name)
-            if ext == '.skl':
-                import_skl_file(fpath, import_context)
-            elif ext == '.skls':
-                import_skls_file(fpath, import_context)
-            else:
-                self.report({'ERROR'}, 'Format of {} not recognised'.format(file))
-        return {'FINISHED'}
-
-    @invoke_require_armature
-    def invoke(self, context, event):
-        return super().invoke(context, event)
-
-
-class ModelExportHelper:
-    selection_only = bpy.props.BoolProperty(
-        name='Selection Only',
-        description='Export only selected objects'
-    )
-
-    def export(self, bpy_obj, context):
-        pass
-
-    @execute_with_logger
-    @execute_require_filepath
-    def execute(self, context):
-        objs = context.selected_objects if self.selection_only else context.scene.objects
-        roots = [obj for obj in objs if obj.xray.isroot]
-        if not roots:
-            self.report({'ERROR'}, 'Cannot find object root')
-            return {'CANCELLED'}
-        if len(roots) > 1:
-            self.report({'ERROR'}, 'Too many object roots found')
-            return {'CANCELLED'}
-        return self.export(roots[0], context)
-
-
-def mk_export_context(texname_from_path, fmt_version=None, export_motions=True):
-    from .obj.exp import ExportContext
-    return ExportContext(
-        textures_folder=plugin_prefs.get_preferences().textures_folder_auto,
-        export_motions=export_motions,
-        soc_sgroups=None if fmt_version is None else (fmt_version == 'soc'),
-        texname_from_path=texname_from_path
-    )
-
-
-@registry.module_thing
-class OpExportOgf(bpy.types.Operator, io_utils.ExportHelper, ModelExportHelper):
-    bl_idname = 'xray_export.ogf'
-    bl_label = 'Export .ogf'
-
-    filename_ext = '.ogf'
-    filter_glob = bpy.props.StringProperty(default='*'+filename_ext, options={'HIDDEN'})
-
-    texture_name_from_image_path = plugin_prefs.PropObjectTextureNamesFromPath()
-
-    def export(self, bpy_obj, context):
-        from .ogf import exp
-        export_context = mk_export_context(self.texture_name_from_image_path)
-        exp.export_file(bpy_obj, self.filepath, export_context)
-        return {'FINISHED'}
-
-    def invoke(self, context, event):
-        prefs = plugin_prefs.get_preferences()
-        self.texture_name_from_image_path = prefs.object_texture_names_from_path
-        return super().invoke(context, event)
-
-
-@registry.module_thing
-class OpExportSkl(bpy.types.Operator, io_utils.ExportHelper):
-    bl_idname = 'xray_export.skl'
-    bl_label = 'Export .skl'
-    bl_description = 'Exports X-Ray skeletal animation'
-
-    filename_ext = '.skl'
-    filter_glob = bpy.props.StringProperty(default='*' + filename_ext, options={'HIDDEN'})
-    action = None
-
-    @execute_with_logger
-    @execute_require_filepath
-    def execute(self, context):
-        from .skl.exp import export_skl_file, ExportContext
-        export_context = ExportContext(
-            armature=context.active_object,
-            action=self.action
-        )
-        export_skl_file(self.filepath, export_context)
-        return {'FINISHED'}
-
-    @invoke_require_armature
-    def invoke(self, context, event):
-        self.action = getattr(context, OpExportSkl.bl_idname + '.action', None)
-        assert self.action
-        self.filepath = self.action.name
-        if not self.filepath.lower().endswith(self.filename_ext):
-            self.filepath += self.filename_ext
-        return super().invoke(context, event)
-
-
-@registry.module_thing
-class OpExportSkls(bpy.types.Operator, FilenameExtHelper):
-    bl_idname = 'xray_export.skls'
-    bl_label = 'Export .skls'
-    bl_description = 'Exports X-Ray skeletal animation'
-
-    filename_ext = '.skls'
-    filter_glob = bpy.props.StringProperty(default='*' + filename_ext, options={'HIDDEN'})
-
-    def export(self, context):
-        from .skl.exp import export_skls_file, ExportContext
-        export_context = ExportContext(
-            armature=context.active_object
-        )
-        export_skls_file(self.filepath, export_context)
-
-    @invoke_require_armature
-    def invoke(self, context, event):
-        return super().invoke(context, event)
-
-
-@registry.module_thing
 class OpExportProject(TestReadyOperator):
     bl_idname = 'export_scene.xray'
     bl_label = 'Export XRay Project'
@@ -372,7 +147,7 @@ class XRayImportMenu(bpy.types.Menu):
             text='Source Object (.object)'
         )
         layout.operator(anm_ops.OpImportAnm.bl_idname, text='Animation (.anm)')
-        layout.operator(OpImportSkl.bl_idname, text='Skeletal Animation (.skl, .skls)')
+        layout.operator(skl_ops.OpImportSkl.bl_idname, text='Skeletal Animation (.skl, .skls)')
         layout.operator(det_ops.OpImportDM.bl_idname, text='Details (.dm, .details)')
         layout.operator(err_ops.OpImportERR.bl_idname, text='Error List (.err)')
 
@@ -390,8 +165,8 @@ class XRayExportMenu(bpy.types.Menu):
             text='Source Object (.object)'
         )
         layout.operator(anm_ops.OpExportAnm.bl_idname, text='Animation (.anm)')
-        layout.operator(OpExportSkls.bl_idname, text='Skeletal Animation (.skls)')
-        layout.operator(OpExportOgf.bl_idname, text='Game Object (.ogf)')
+        layout.operator(skl_ops.OpExportSkls.bl_idname, text='Skeletal Animation (.skls)')
+        layout.operator(ogf_ops.OpExportOgf.bl_idname, text='Game Object (.ogf)')
         layout.operator(det_ops.OpExportDMs.bl_idname, text='Detail Model (.dm)')
         layout.operator(
             det_ops.OpExportLevelDetails.bl_idname,
@@ -439,7 +214,7 @@ def menu_func_import(self, _context):
         icon_value=icon
     )
     self.layout.operator(anm_ops.OpImportAnm.bl_idname, text='X-Ray animation (.anm)', icon_value=icon)
-    self.layout.operator(OpImportSkl.bl_idname, text='X-Ray skeletal animation (.skl, .skls)', icon_value=icon)
+    self.layout.operator(skl_ops.OpImportSkl.bl_idname, text='X-Ray skeletal animation (.skl, .skls)', icon_value=icon)
 
 
 def menu_func_export(self, _context):
@@ -450,12 +225,12 @@ def menu_func_export(self, _context):
         icon_value=icon
     )
     self.layout.operator(anm_ops.OpExportAnm.bl_idname, text='X-Ray animation (.anm)', icon_value=icon)
-    self.layout.operator(OpExportSkls.bl_idname, text='X-Ray animation (.skls)', icon_value=icon)
+    self.layout.operator(skl_ops.OpExportSkls.bl_idname, text='X-Ray animation (.skls)', icon_value=icon)
 
 
 def menu_func_export_ogf(self, _context):
     icon = get_stalker_icon()
-    self.layout.operator(OpExportOgf.bl_idname, text='X-Ray game object (.ogf)', icon_value=icon)
+    self.layout.operator(ogf_ops.OpExportOgf.bl_idname, text='X-Ray game object (.ogf)', icon_value=icon)
 
 
 def menu_func_xray_import(self, _context):
@@ -518,6 +293,8 @@ def register():
     registry.register_thing(object_imp_ops, __name__)
     registry.register_thing(object_exp_ops, __name__)
     registry.register_thing(anm_ops, __name__)
+    registry.register_thing(skl_ops, __name__)
+    registry.register_thing(ogf_ops, __name__)
     scene_ops.register_operators()
     det_ops.register_operators()
     registry.register_thing(err_ops, __name__)
@@ -534,6 +311,8 @@ def unregister():
     registry.unregister_thing(err_ops, __name__)
     det_ops.unregister_operators()
     scene_ops.unregister_operators()
+    registry.unregister_thing(ogf_ops, __name__)
+    registry.unregister_thing(skl_ops, __name__)
     registry.unregister_thing(anm_ops, __name__)
     registry.unregister_thing(object_exp_ops, __name__)
     registry.unregister_thing(object_imp_ops, __name__)
